% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_guard_additive.R
\name{shift_additive_guard}
\alias{shift_additive_guard}
\alias{shift_additive_guard_inv}
\alias{get_bounds_maximal}
\title{Additive Shifts of Continuous-Valued Interventions with Bound Guards}
\usage{
shift_additive_guard(tmle_task, delta, likelihood_base, ...)

shift_additive_guard_inv(tmle_task, delta, likelihood_base, ...)

get_bounds_maximal(tmle_task, likelihood_base)
}
\arguments{
\item{tmle_task}{A \code{tmle3_Task} object containing data and nodes, as
described and implemented in the \code{tmle3} package. Please refer to the
documentation and supporting materials of that package for details.}

\item{delta}{A \code{numeric} value giving a value of the shift to be applied
to the treatment. This is an additive shift so the value is merely to be
added to the observed value of the treatment node "A". In the case of the
inverse additive shift, the specified value will be subtracted from the
observed value of the treatment node "A".}

\item{likelihood_base}{The base observed data likelihood, to be used in
implementing guards that ensure that the shifted treatment does not violate
the bounds induced by the support of the intervention, conditional on the
covariates.}

\item{...}{Additional arguments (currently unused).}

\item{tmle_task}{A \code{tmle3_Task} object containing data and nodes, as
described and implemented in the \code{tmle3} package. Please refer to the
documentation and supporting materials of that package for details.}

\item{likelihood_base}{The base observed data likelihood, to be used in
implementing guards that ensure that the shifted treatment does not violate
the bounds induced by the support of the intervention, conditional on the
covariates.}
}
\description{
Additive Shifts of Continuous-Valued Interventions with Bound Guards

Find Maximum Plausible Intervention Shifts from the Observed Data
}
\seealso{
Other shifting_interventions: \code{\link{shift_additive_bounded}},
  \code{\link{shift_additive}}

Other shifting_interventions: \code{\link{shift_additive_bounded}},
  \code{\link{shift_additive}}

Other shifting_interventions: \code{\link{shift_additive_bounded}},
  \code{\link{shift_additive}}
}
\concept{shifting_interventions}
\keyword{internal}
