% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_msm_functions.R
\docType{data}
\name{delta_param_MSM_linear}
\alias{delta_param_MSM_linear}
\title{Linear Working Marginal Structural Models}
\format{An object of class \code{list} of length 4.}
\usage{
delta_param_MSM_linear
}
\description{
Linear Working Marginal Structural Models
}
\keyword{datasets}
