% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_shift.R
\docType{data}
\name{LF_shift}
\alias{LF_shift}
\title{Shifted Likelihood Factor}
\format{\code{\link{R6Class}} object.}
\usage{
LF_shift
}
\value{
\code{LF_base} object
}
\description{
Shifts a likelihood factor according to a \code{shift_function} and a given
magnitude of the desired shift (\code{shift_delta}). In effect,
\code{get_likelihood(tmle_task)} from \code{tmle3} will instead be the
likelihood from the \code{original_lf}, but a for shifted value
\eqn{A'=}\code{shift_function}\eqn{(A, W)}
}
\section{Constructor}{

  \code{define_lf(LF_shift, name, type = "density", original_lf,
    shift_function, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node
          name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{original_lf}}{\code{\link{LF_base}} object, the likelihood
          factor to shift
    }
    \item{\code{shift_function}}{\code{function}, defines the shift
    }
    \item{\code{shift_inverse}}{\code{function}, the inverse of a given
          \code{shift_function}
    }
    \item{\code{shift_delta}}{\code{numeric}, specification of the magnitude
          of the desired shift (on the level of the treatment)
    }
    \item{\code{max_shifted_ratio}}{A \code{numeric} value indicating the maximum
          tolerance for the ratio of the counterfactual and observed
          intervention densities. In particular, the shifted value of the
          intervention is assigned to a given observational unit when the
          ratio of the counterfactual intervention density to the observed
          intervention density is below this value
   }
    \item{\code{...}}{Not currently used.
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{original_lf}}{\code{\link{LF_base}} object, the likelihood
          factor to shift
    }
    \item{\code{shift_function}}{\code{function}, defines the shift
    }
    \item{\code{shift_inverse}}{\code{function}, the inverse of a given
          \code{shift_function}
    }
    \item{\code{shift_delta}}{\code{numeric}, specification of the magnitude
          of the desired shift (on the level of the treatment)
   }
    \item{\code{max_shifted_ratio}}{A \code{numeric} value indicating the maximum
          tolerance for the ratio of the counterfactual and observed
          intervention densities. In particular, the shifted value of the
          intervention is assigned to a given observational unit when the
          ratio of the counterfactual intervention density to the observed
          intervention density is below this value
   }
    \item{\code{...}}{Not currently used.
    }
  }
}

\references{
\describe{
  \item{"Stochastic Treatment Regimes."}{Díaz, Iván and van der Laan, Mark J
        (2018). In Targeted Learning in Data Science: Causal Inference for
        Complex Longitudinal Studies, 167–80. Springer Science & Business
        Media.}
  \item{"Population Intervention Causal Effects Based on Stochastic
        Interventions."}{Díaz, Iván and van der Laan, Mark J (2012).
        Biometrics 68 (2). Wiley Online Library: 541–49.}
}
}
\concept{Likelihood objects}
\keyword{data}
