% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_MSM_linear.R
\docType{data}
\name{Param_MSM_linear}
\alias{Param_MSM_linear}
\title{Parameter for Linear Working Marginal Structural Model}
\format{\code{\link{R6Class}} object.}
\usage{
Param_MSM_linear
}
\value{
\code{Param_base} object
}
\description{
Parameter definition for targeting the parameters of a linear working
marginal structural model (MSM): EY = beta0 + beta1 * delta, in order to
summarize the variable importance results of a grid of shift interventions.
}
\section{Constructor}{

  \code{define_param(Param_MSM_linear, observed_likelihood,
                     intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}}
      corresponding to the observed likelihood.}
    \item{\code{intervention_list}}{A list of objects inheriting from
      \code{\link{LF_base}}, representing the intervention.}
    \item{\code{...}}{Not currently used.}
    \item{\code{outcome_node}}{character, the name of the node that should
      be treated as the outcome.}
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this
      treatment.}
    \item{\code{intervention_list}}{A list of objects inheriting from
      \code{\link{LF_base}}, representing the intervention.}
}
}

\concept{Parameters}
\keyword{data}
