% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_shift.R
\docType{data}
\name{tmle3_Spec_shift}
\alias{tmle3_Spec_shift}
\title{Defines a TML Estimator for the Outcome under a Delta-Shifted Treatment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_shift
}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}
See TODO notes for places generalization can be added
}
\keyword{datasets}
