% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_vimshift_msm.R
\name{tmle3_Spec_vimshift_msm}
\alias{tmle3_Spec_vimshift_msm}
\title{Defines a TML Estimator for Variable Importance for Continuous Interventions}
\description{
Defines a TML Estimator for Variable Importance for Continuous Interventions
}
