% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_shift.R
\name{tmle_shift}
\alias{tmle_shift}
\title{Outcome under Shifted Treatment}
\usage{
tmle_shift(shift_fxn = shift_additive,
  shift_fxn_inv = shift_additive_inv, shift_val = 0)
}
\arguments{
\item{shift_fxn}{A \code{function} defining the type of shift to be applied
to the treatment. For an example, see \code{shift_additive}.}

\item{shift_fxn_inv}{A \code{function} defining the inverse of the type of
shift to be applied to the treatment. For an example, see
\code{shift_additive_inv}.}

\item{shift_val}{A \code{numeric}, specification of the magnitude of the
desired shift (on the level of the treatment). This is a value passed to
the \code{function}s above for modulating the treatment.}
}
\description{
O = (W, A, Y)
W = Covariates
A = Treatment (binary or categorical)
Y = Outcome (binary or bounded continuous)
}
