% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_vimshift_delta.R
\name{tmle_vimshift_delta}
\alias{tmle_vimshift_delta}
\title{Outcome Under a Grid of Shifted Interventions via Delta Method}
\usage{
tmle_vimshift_delta(
  shift_fxn = shift_additive_bounded,
  shift_fxn_inv = shift_additive_bounded_inv,
  shift_grid = seq(-1, 1, by = 0.5),
  max_shifted_ratio = 2,
  weighting = c("identity", "variance"),
  ...
)
}
\arguments{
\item{shift_fxn}{A \code{function} defining the type of shift to be applied
to the treatment. For an example, see \code{shift_additive}.}

\item{shift_fxn_inv}{A \code{function} defining the inverse of the type of
shift to be applied to the treatment. For an example, see
\code{shift_additive_inv}.}

\item{shift_grid}{A \code{numeric} vector, specification of a selection of
shifts (on the level of the treatment) to be applied to the intervention.
This is a value passed to the \code{function}s above for computing various
values of the outcome under modulated values of the treatment.}

\item{max_shifted_ratio}{A \code{numeric} value indicating the maximum
tolerance for the ratio of the counterfactual and observed intervention
densities. In particular, the shifted value of the intervention is assigned
to a given observational unit when the ratio of counterfactual intervention
density to the observed intervention density is below this value.}

\item{weighting}{A \code{character} indicating the type of weighting used
for construction of the marginal structural model. \code{"identity"}
applies the same weight to all individual estimates while \code{"variance"}
applies weights based on the inverse variance of the estimate. It would be
expected that variance-based weighting would yield more stable estimates of
the parameter of the MSM. The default remains the identity weighting.}

\item{...}{Additional arguments, passed to shift functions.}
}
\description{
O = (W, A, Y)
W = Covariates
A = Treatment (binary or categorical)
Y = Outcome (binary or bounded continuous)
}
