% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_test.R
\name{trend_msm}
\alias{trend_msm}
\title{Test for a trend in the effect of shift interventions}
\usage{
trend_msm(tmle_fit, delta_grid, level = 0.95, weights = rep(1,
  length(delta_grid)))
}
\arguments{
\item{tmle_fit}{An object of class \code{tmle3_Fit}, containing estimates of
a grid of posited shift interventions.}

\item{delta_grid}{A \code{numeric} vector giving the individual values of the
shift parameter used in computing each of the TML estimates.}

\item{level}{The nominal coverage probability of the confidence interval.}

\item{weights}{A \code{numeric} vector indicating the weights (if any) to be
applied to each of the posited values of the shift parameter delta. The
default is equal weighting across all shift values.}
}
\description{
Test for a trend in the effect of shift interventions
}
