% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_additive.R
\name{shift_additive}
\alias{shift_additive}
\alias{shift_additive_inv}
\title{Additive Shifts of Continuous-Valued Interventions}
\usage{
shift_additive(tmle_task, delta = 0, ...)

shift_additive_inv(tmle_task, delta = 0, ...)
}
\arguments{
\item{tmle_task}{A \code{tmle3_Task} object containing data and nodes, as
described and implemented in the \code{tmle3} package. Please refer to the
documentation and supporting materials of that package for details.}

\item{delta}{A \code{numeric} value giving a value of the shift to be applied
to the treatment. This is an additive shift so the value is merely to be
added to the observed value of the treatment node "A". In the case of the
inverse additive shift, the specified value will be subtracted from the
observed value of the treatment node "A".}

\item{...}{Additional arguments (currently unused).}
}
\description{
Additive Shifts of Continuous-Valued Interventions
}
\seealso{
Other shifting_interventions: \code{\link{shift_additive_guard}}

Other shifting_interventions: \code{\link{shift_additive_guard}}
}
\concept{shifting_interventions}
