% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_msm_functions.R
\name{msm_linear_factory}
\alias{msm_linear_factory}
\title{Function factory for computing linear working MSM delta method functions}
\usage{
msm_linear_factory(design_matrix)
}
\arguments{
\item{design_matrix}{A \code{matrix} or \code{data.frame} giving the design
matrix to be used in specifying terms of the linear working marginal
structural model (MSM) for which parameters are to be estimated.}
}
\description{
Function factory for computing linear working MSM delta method functions
}
\keyword{internal}
