% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_shift.R
\name{tmle3_Spec_shift}
\alias{tmle3_Spec_shift}
\title{Defines a TML Estimator for the Outcome under a Shifted Treatment}
\description{
Defines a TML Estimator for the Outcome under a Shifted Treatment
}
