% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_vimshift_delta.R
\docType{data}
\name{tmle3_Spec_vimshift_delta}
\alias{tmle3_Spec_vimshift_delta}
\title{Defines a TML Estimator for Variable Importance for Continuous Interventions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_vimshift_delta
}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}
See TODO notes for places generalization can be added
}
\keyword{datasets}
