% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_vimshift_msm.R
\docType{data}
\name{tmle3_Spec_vimshift_msm}
\alias{tmle3_Spec_vimshift_msm}
\title{Defines a TML Estimator for Variable Importance for Continuous Interventions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_vimshift_msm
}
\description{
Current limitations: pretty much tailored to \code{Param_TSM}.
}
\keyword{datasets}
