% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_test.R
\name{trend_msm}
\alias{trend_msm}
\title{Test for a trend in the effect of shift interventions}
\usage{
trend_msm(tmle_fit, delta_grid, level = 0.95, weights = NULL)
}
\arguments{
\item{tmle_fit}{An object of class \code{tmle3_Fit}, containing estimates of
a grid of posited shift interventions.}

\item{delta_grid}{A \code{numeric} vector giving the individual values of the
shift parameter used in computing each of the TML estimates.}

\item{level}{The nominal coverage probability of the confidence interval.}

\item{weights}{A \code{numeric} vector indicating the weights (if any) to be
applied to each of the estimated mean counterfactual outcomes under posited
values of the shift parameter delta. The default is to weight each estimate
by the inverse of its variance, in order to improve stability; however, this
may be changed depending on the exact choice of shift function.}
}
\description{
Test for a trend in the effect of shift interventions
}
